requirements.txt
# streamlit
# langchain
# langchain_groq
# faiss-cpu
# huggingface-hub
# transformers
# sentence-transformers
# python-dotenv
# PyPDF2

streamlit
langchain-groq
python-dotenv
pypdf==5.0.0
faiss-cpu
huggingface-hub
langchain
langchain-community
python-docx